Rename project from github-ci-fetcher to cimonitor
Complete rebranding to a more LLM-friendly name:
- Package: github-ci-fetcher → cimonitor
- CLI command: fetch-ci-logs → cimonitor
- Update all import references and documentation
- Update pyproject.toml, README.md, mise.toml, CI workflows
- Clean description: "Monitor GitHub CI workflows, fetch logs, and track build status"

The name "cimonitor" clearly indicates to LLMs that this tool can:
- Monitor CI status and workflows
- Fetch logs and track build status
- Provide real-time CI monitoring capabilities
