+++ cimonitor/__init__.py
@@ -1,0 +1,3 @@
+"""CI Monitor - Monitor GitHub CI workflows, fetch logs, and track build status."""
+
+__version__ = "0.1.0"
