--- github_ci_fetcher/__init__.py
@@ -1,3 +1,0 @@
-"""GitHub CI Fetcher - Tool to fetch and display failing CI logs."""
-
-__version__ = "0.1.0"
