+++ pyproject.toml
@@ -1,13 +1,13 @@
 [project]
-name = "github-ci-fetcher"
+name = "cimonitor"
 version = "0.1.0"
-description = "Tool to fetch and display GitHub CI logs for failing builds"
+description = "Monitor GitHub CI workflows, fetch logs, and track build status"
 readme = "README.md"
 license = { text = "MIT" }
 authors = [
     { name = "Steve Landey", email = "steve@steveasleep.com" }
 ]
-keywords = ["github", "ci", "logs", "actions", "debugging", "devops"]
+keywords = ["ci", "monitor", "github", "logs", "actions", "debugging", "devops", "workflow"]
 classifiers = [
     "Development Status :: 4 - Beta",
     "Intended Audience :: Developers",
@@ -34,7 +34,7 @@ dependencies = [
 ]
 
 [project.scripts]
-fetch-ci-logs = "github_ci_fetcher.cli:main"
+cimonitor = "cimonitor.cli:main"
 
 [project.optional-dependencies]
 dev = [
